Split everything into it's own files, added unit tests for sliders.
